requirements.txt
langgraph
langchain
langchain-openai
azure-storage-blob
python-dotenv
